// src/app/model/login.ts
export class Login {
  action: string | undefined;
  email: string | undefined;
  password: string | undefined;
}

export class Loginresponse {
  id: Number | undefined;
  type: string | undefined;
  status: string | undefined;
  message: string | undefined;
  first_name: string | undefined;
  last_name: string | undefined;
  register_no: Number | undefined;
  user_role: Number | undefined;
  skill: Number | undefined;
}